+++ packages/scanners/handshake-scanner/lib/scanner/handshakeRpcScanner.ts
@@ -1,0 +1,18 @@
+import { GeneralScanner, ScannerConfig } from '@rosen-bridge/abstract-scanner';
+
+import { HandshakeRpcTransaction } from '../types';
+
+export class HandshakeRpcScanner extends GeneralScanner<HandshakeRpcTransaction> {
+  constructor(config: ScannerConfig<HandshakeRpcTransaction>) {
+    super(
+      'handshake',
+      config.dataSource,
+      config.initialHeight,
+      config.network,
+      config.blockRetrieveGap,
+      config.blockTimeConfig,
+      config.logger,
+      config.suffix,
+    );
+  }
+}