+++ packages/abstract-observation-extractor/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @rosen-bridge/abstract-observation-extractor
 
+## 1.0.6
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/rosen-extractor@12.0.2
+
+## 1.0.5
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.2
+  - @rosen-bridge/extended-typeorm@1.1.0
+  - @rosen-bridge/rosen-extractor@12.0.1
+  - @rosen-bridge/tokens@6.0.1
+
+## 1.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.1
+
+## 1.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.0
+  - @rosen-bridge/rosen-extractor@11.3.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Update dependencies:
+  - @rosen-bridge/rosen-extractor@11.2.2
+  - @rosen-bridge/tokens@6.0.0
+
 ## 1.0.1
 
 ### Patch Changes