+++ packages/abstract-scanner/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rosen-bridge/abstract-scanner
 
+## 1.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.2
+  - @rosen-bridge/extended-typeorm@1.1.0
+
+## 1.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.1
+
+## 1.0.2
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.0
+
 ## 1.0.1
 
 ### Patch Changes