+++ packages/evm-address-tx-extractor/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @rosen-bridge/evm-address-tx-extractor
 
+## 2.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.2
+  - @rosen-bridge/extended-typeorm@1.1.0
+
+## 2.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.1
+
+## 2.0.2
+
+### Patch Changes
+
+- Fix typo in logs descriptions
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.0
+
 ## 2.0.1
 
 ### Patch Changes