+++ packages/observation-extractors/bitcoin-observation-extractor/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @rosen-bridge/bitcoin-observation-extractor
 
+## 8.0.1
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.6
+  - @rosen-bridge/rosen-extractor@12.0.2
+
+## 8.0.0
+
+### Major Changes
+
+- Any service that uses this packages requires to initialize the `AddressManager`, provided by `@rosen-bridge/address-manager`, so the Rosen extractor becomes functional
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.5
+  - @rosen-bridge/bitcoin-scanner@1.0.4
+  - @rosen-bridge/extended-typeorm@1.1.0
+  - @rosen-bridge/rosen-extractor@12.0.1
+  - @rosen-bridge/tokens@6.0.1
+
+## 7.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.4
+  - @rosen-bridge/bitcoin-scanner@1.0.3
+
+## 7.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@1.0.3
+  - @rosen-bridge/bitcoin-scanner@1.0.2
+  - @rosen-bridge/rosen-extractor@11.3.0
+
+## 7.0.2
+
+### Patch Changes
+
+- Update dependencies:
+  - @rosen-bridge/rosen-extractor@11.2.2
+  - @rosen-bridge/tokens@6.0.0
+  - @rosen-bridge/abstract-observation-extractor@1.0.2
+
 ## 7.0.1
 
 ### Patch Changes