+++ packages/scanners/bitcoin-scanner/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @rosen-bridge/bitcoin-scanner
 
+## 1.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.3
+  - @rosen-clients/rate-limited-axios@2.0.0
+
+## 1.0.2
+
+### Patch Changes
+
+- @rosen-bridge/abstract-scanner@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes