+++ packages/scanners/cardano-scanner/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rosen-bridge/cardano-scanner
 
+## 2.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.4
+  - @rosen-bridge/extended-typeorm@1.1.0
+
+## 2.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.3
+  - @rosen-clients/rate-limited-axios@2.0.0
+
+## 2.0.2
+
+### Patch Changes
+
+- @rosen-bridge/abstract-scanner@1.0.2
+
 ## 2.0.1
 
 ### Patch Changes