+++ packages/scanners/evm-scanner/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rosen-bridge/evm-scanner
 
+## 1.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @rosen-bridge/abstract-scanner@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes