+++ packages/tx-id-extractor/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rosen-bridge/tx-id-extractor
 
+## 3.0.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.2
+  - @rosen-bridge/extended-typeorm@1.1.0
+
+## 3.0.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.1
+
+## 3.0.2
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.1.0
+
 ## 3.0.1
 
 ### Patch Changes